+++ packages/admin/admin-stories/src/docs/form/components/stories/FinalFormFields.stories.tsx
@@ -77,6 +77,7 @@ storiesOf("stories/form/FinalForm Fields", module)
             () => ({
                 autocomplete: { value: "strawberry", label: "Strawberry" },
                 autocompleteAsync: { value: "strawberry", label: "Strawberry" },
+                autocompleteMultiple: [{ value: "strawberry", label: "Strawberry" }],
             }),
             [],
         );
@@ -96,9 +97,7 @@ storiesOf("stories/form/FinalForm Fields", module)
                 <Field
                     component={FinalFormAutocomplete}
                     getOptionLabel={(option: Option) => option.label}
-                    getOptionSelected={(option: Option, value: Option) => {
-                        return option.value === value.value;
-                    }}
+                    isOptionEqualToValue={(option: Option, value: Option) => option.value === value.value}
                     options={options}
                     name="autocomplete"
                     label="Autocomplete"
@@ -108,11 +107,19 @@ storiesOf("stories/form/FinalForm Fields", module)
                     component={FinalFormAutocomplete}
                     {...acAsyncProps}
                     getOptionLabel={(option: Option) => option.label}
-                    getOptionSelected={(option: Option, value: Option) => {
-                        return option.value === value.value;
-                    }}
+                    isOptionEqualToValue={(option: Option, value: Option) => option.value === value.value}
                     name="autocompleteAsync"
                     label="AutocompleteAsync"
+                    fullWidth
+                />
+                <Field
+                    component={FinalFormAutocomplete}
+                    multiple
+                    getOptionLabel={(option: Option) => option.label}
+                    isOptionEqualToValue={(option: Option, value: Option) => option.value === value.value}
+                    options={options}
+                    name="autocompleteMultiple"
+                    label="Autocomplete multiple select"
                     fullWidth
                 />
                 <Button color="primary" variant="contained" type="submit">