+++ packages/admin/admin-theme/src/componentsTheme/getComponentsTheme.ts
@@ -40,6 +40,7 @@ import { getMuiTableRow } from "./MuiTableRow";
 import { getMuiTabs } from "./MuiTabs";
 import { getMuiToggleButton } from "./MuiToggleButton";
 import { getMuiToggleButtonGroup } from "./MuiToggleButtonGroup";
+import { getMuiTooltip } from "./MuiTooltip";
 import { getMuiTypography } from "./MuiTypography";
 
 type ThemeData = {
@@ -92,5 +93,6 @@ export const getComponentsTheme = (components: Components, themeData: ThemeData)
     MuiTabs: getMuiTabs(components.MuiTabs, themeData),
     MuiToggleButton: getMuiToggleButton(components.MuiToggleButton, themeData),
     MuiToggleButtonGroup: getMuiToggleButtonGroup(components.MuiToggleButtonGroup, themeData),
+    MuiTooltip: getMuiTooltip(components.MuiTooltip, themeData),
     MuiTypography: getMuiTypography(components.MuiTypography, themeData),
 });