+++ packages/api/cms-api/src/builds/dto/build-template.object.ts
@@ -1,5 +1,7 @@
 import { Field, ID, ObjectType } from "@nestjs/graphql";
 
+import { LABEL_ANNOTATION } from "../builds.constants";
+
 @ObjectType("BuildTemplate")
 export class BuildTemplateObject {
     @Field(() => ID)
@@ -7,4 +9,7 @@ export class BuildTemplateObject {
 
     @Field()
     name: string;
+
+    @Field({ nullable: true, description: `Human readable label provided by ${LABEL_ANNOTATION} annotation. Use name as fallback if not present` })
+    label?: string;
 }