+++ packages/admin/admin-icons/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @comet/admin-icons
 
+## 4.4.0
+
+## 4.3.0
+
+## 4.2.0
+
 ## 4.1.0
 
 ### Minor Changes