+++ packages/admin/admin-react-select/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @comet/admin-react-select
 
+## 4.4.0
+
+### Patch Changes
+
+-   Updated dependencies [e824ffa6]
+-   Updated dependencies [3e15b819]
+-   Updated dependencies [a77da844]
+    -   @comet/admin@4.4.0
+
+## 4.3.0
+
+### Patch Changes
+
+-   @comet/admin@4.3.0
+
+## 4.2.0
+
+### Patch Changes
+
+-   Updated dependencies [67e54a82]
+-   Updated dependencies [3567533e]
+-   Updated dependencies [7b614c13]
+-   Updated dependencies [aaf1586c]
+-   Updated dependencies [d25a7cbb]
+    -   @comet/admin@4.2.0
+
 ## 4.1.0
 
 ### Patch Changes