+++ packages/admin/admin-theme/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @comet/admin-theme
 
+## 4.4.0
+
+### Patch Changes
+
+-   @comet/admin-icons@4.4.0
+
+## 4.3.0
+
+### Patch Changes
+
+-   @comet/admin-icons@4.3.0
+
+## 4.2.0
+
+### Patch Changes
+
+-   67e54a82: Add styling variants to Tooltip
+    -   @comet/admin-icons@4.2.0
+
 ## 4.1.0
 
 ### Patch Changes