+++ packages/admin/blocks-admin/CHANGELOG.md
@@ -1,5 +1,40 @@
 # @comet/blocks-admin
 
+## 4.4.0
+
+### Minor Changes
+
+-   d4960b05: Add loop toggle to YouTubeVideo block
+
+### Patch Changes
+
+-   Updated dependencies [e824ffa6]
+-   Updated dependencies [3e15b819]
+-   Updated dependencies [a77da844]
+    -   @comet/admin@4.4.0
+    -   @comet/admin-icons@4.4.0
+
+## 4.3.0
+
+### Patch Changes
+
+-   @comet/admin@4.3.0
+-   @comet/admin-icons@4.3.0
+
+## 4.2.0
+
+### Patch Changes
+
+-   b4d564b6: Add spacing by default between the tabs and the tab content when using `AdminTabs` or `BlockPreviewWithTabs`. This may add unwanted additional spacing in cases where the spacing was added manually.
+-   0f4ed6c1: Prevent unintentional movement of sticky header when scrolling in Tabs.
+-   Updated dependencies [67e54a82]
+-   Updated dependencies [3567533e]
+-   Updated dependencies [7b614c13]
+-   Updated dependencies [aaf1586c]
+-   Updated dependencies [d25a7cbb]
+    -   @comet/admin@4.2.0
+    -   @comet/admin-icons@4.2.0
+
 ## 4.1.0
 
 ### Patch Changes