+++ packages/api/blocks-api/CHANGELOG.md
@@ -1,3 +1,13 @@
 # @comet/blocks-api
 
+## 4.4.0
+
+### Minor Changes
+
+-   d4960b05: Add loop toggle to YouTubeVideo block
+
+## 4.3.0
+
+## 4.2.0
+
 ## 4.1.0